requirements.txt
amqp==2.1.4
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
bleach==1.5.0
celery==4.0.2
certifi==2016.9.26
configparser==3.5.0
decorator==4.0.10
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
drfdocs==0.0.11
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
html5lib==0.9999999
ijson==2.3
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
kombu==4.0.2
Markdown==2.6.7
MarkupSafe==0.23
mistune==0.7.3
nbconvert==5.0.0
nbformat==4.2.0
notebook==4.3.1
numpy==1.11.3
olefile==0.44
pandas==0.19.2
pandocfilters==1.4.1
pathlib2==2.2.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.2.1
redis==2.10.5
requests==2.12.4
scandir==1.4
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
widgetsnbextension==1.2.6
